pyproject.toml
[tool.poetry]
name = "barcodes recognizer via telegram"
version = "0.1.0"
description = "This is a barcodes (including ean, qr and other datamatrix codes) recognizer with chatbot interface for Telegram."
authors = ["benzlokzik"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Pillow = "9.4.0"
python-barcode = "^0.14.0"
python-telegram-bot = "20.0"
pyzbar = "^0.1.9"
qrcode = "^7.3.1"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
flake8 = "==6.0.0"
pyright = "==1.1.291"
#requests = "^2.28.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
